# [BugFix] Use async Mistral Tokenizer in Chat Completions