## Purpose

When using `--tokenizer-mode mistral` with Chat Completions, large inputs could block the event loop, causing delays in processing of other requests. This adjusts the usage of that tokenizer during chat request preprocessing to use pre-existing but unused `_tokenizer_executor` in OpenAIServing to run these blocking operations in a background thread as opposed to running them directly in the critical path of the server event loop.

Fixes #24910 

## Test Plan

I added a new `tests/entrypoints/openai/test_serving_engine.py` with a single test that ensures the newly extracted `_apply_chat_template` method properly offloads the long-running tokenization operation to a background thread and does not block the event loop. For testing simplicity, this mocks the actual tokenization to just sleep to get a consistent simulation of a long tokenization.

To run:

```
 pytest -sv tests/entrypoints/openai/test_serving_engine.py
```

## Test Result

I ensured the test failed before I moved the execution into the threadpool executor. And, after moving it into the executor, the test passes. The test doesn't count the event loop as blocked until it's been blocked for at least 0.5 seconds, which is far higher than required in most scenarios but a value that large was chosen to ensure this doesn't flake on a heavily loaded or slow CI machine.