@bbrowning 2025-10-02
For a visual overview of the changes here, below is a before and after sequence diagram. This does not change the time a single large request takes when using `--tokenizer mistral`. If that took 50 seconds before, it will still take 50 seconds. All this does is free up the web server's event loop for processing other requests during that 50 seconds, which allows things like requests to `/health` or other APIs to respond quickly. If two large input requests come in that each take 50 seconds, those would take 100 seconds to complete before and will still take 100 seconds to complete now.

In other words, this isn't adding any parallelization of the Mistral tokenizer usage. That could be something considered as a separate PR depending on community and user needs for concurrent request performance with large inference inputs with that tokenizer.

## Before

```mermaid
  sequenceDiagram
      participant U1 as User 1
      participant U2 as User 2 (Health Check)
      participant EL as Event Loop
      participant MT as Mistral Tokenizer

      Note over U1,MT: BEFORE: Blocking Tokenization
      U1->>+EL: POST /v1/chat/completions (large input)
      EL->>MT: apply_mistral_chat_template
      activate MT
      Note over MT: 50 seconds processing...

      U2->>EL: GET /health
      Note over EL,U2: Request queued, waiting for event loop

      MT-->>EL: Tokenization complete
      deactivate MT
      EL-->>-U1: Chat Response (after 50s)

      EL-->>U2: Health OK (delayed ~50s)
```

## After

```mermaid
  sequenceDiagram
      participant U1 as User 1
      participant U2 as User 2 (Health Check)
      participant EL as Event Loop
      participant BG as Background Thread
      participant MT as Mistral Tokenizer
      
      Note over U1,BG: AFTER: Async Tokenization
      U1->>+EL: POST /v1/chat/completions (large input)
      EL->>BG: async_apply_mistral_chat_template
      activate BG
      deactivate EL

      BG->>+MT: apply_mistral_chat_template
      Note over MT: 50 seconds processing...

      U2->>+EL: GET /health
      Note over EL: Event loop available!
      EL-->>-U2: Health OK (immediate response)

      MT-->-BG: Tokenization complete
      BG-->>U1: Chat Response (after 50s)
      deactivate BG
```