@bbrowning 2025-10-03
I'm going to mark this as draft, as I see some in-progress work at #22880 that's starting to merge and will impact whether this is still an issue and if so how we fix it. Specifically, the issue described at #23873 as part of that overall work will change this. So, will revisit once some of that lands unless there's a strong desire to preempt that work.

I have also since discovered the existing `make_async` method, so even if we did want to land just this fix I can simplify the implementation.