@bbrowning 2025-10-03
Updated to use `make_async`, pulling out of draft so that others can decide if this should merge sooner or wait on the larger refactoring of chat preprocessing.